+++ src/org/mtransit/android/data/POIArrayAdapter.java
@@ -1249,7 +1249,7 @@ public class POIArrayAdapter extends MTArrayAdapter<POIManager> implements Senso
 				}
 				holder.routeFL.setVisibility(View.VISIBLE);
 				holder.rtsExtraV.setVisibility(View.VISIBLE);
-				final Integer tripId;
+				final Long tripId;
 				if (rts.trip == null) {
 					holder.tripHeadingBg.setVisibility(View.GONE);
 					tripId = null;
